backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
psycopg[binary]==3.2.9
alembic==1.16.4
pydantic==2.11.7
pydantic-settings==2.10.1
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.0
email-validator==2.2.0
redis==6.4.0
httpx==0.28.1
routeros-api==0.21.0
cryptography==45.0.6
python-multipart==0.0.20
